FIXES #205: JS loaded per ajax is not debuggable in Firebug or any other Browser Debugger
- Refactoring the addScript function to not use jQuery.addScript (which
doesn't add a script element in head instead just executes the
script) and instead use plain javascript. The reason is keeping
consistency with the behaviour with tapestry-core and tapestry5-jquery
when SUPPRESS_PROTOTYPE=false. The behaiour was noticed while
integrating the tynamo.org tapestry-ckeditor module with
tapestry5-jquery. CKEditor's initialization javascript looks for
ckeditor.js in the head part in order to set CKEDITOR.basePath to the
base path of the ckeditor folder.