Fixes for broken maven build due to javadoc generation
I went through all files that were printed out in the error log and fixed all of the malformed javadoc comments.
The pom.xml now properly references tapestry’s own javadoc plugin to support @tapestrydoc for component and mixin parameter reports. I removed the annoation from classes that are not in these categories.

Maybe I went a bit out of scope but I also removed the JQueryDateFieldStack since it’s replaced by a contribution made in JQueryModule. Also, maybe backwards incompatible, I moved the abstract base classes of the DataTable component (AbstractQueryTable and AbstractTable) to the *.base package since it is the correct place for abstract classes (as stated in the Tapestry documentation).